+++ tests/Feature/MediaColumnsMustAppearAfterAllOtherFieldsTest.php
@@ -29,7 +29,7 @@ class MediaColumnsMustAppearAfterAllOtherFieldsTest extends TestCase
         'TRANS_TYPE_ID' => '1',
         'BEDROOMS' => '4',
         'PRICE' => '250000',
-        'PRICE_QUALIFIER' => '',
+        'PRICE_QUALIFIER' => '2', // 2 – Guide Price,
         'PROP_SUB_ID' => '0',
         'ADDRESS_1' => 'SZ ADF TestingEstate Agency. ZG Test (ML)',
         'ADDRESS_2' => 'Snowdon Drive, Winterhill',
@@ -39,6 +39,7 @@ class MediaColumnsMustAppearAfterAllOtherFieldsTest extends TestCase
         'FEATURE1' => 'House',
         'FEATURE2' => 'Garden',
         'FEATURE3' => 'Lake',
+        'LET_TYPE_ID' => '0', // 0 = not specified DEFAULT
         'SUMMARY' => 'whatever whatever whatever whatever',
         'DESCRIPTION' => 'whatever whatever whatever whatever whatever whatever',
         'NEW_HOME_FLAG' => '0',
@@ -58,7 +59,7 @@ class MediaColumnsMustAppearAfterAllOtherFieldsTest extends TestCase
             'TRANS_TYPE_ID' => '1',
             'BEDROOMS' => '4',
             'PRICE' => '250000',
-            'PRICE_QUALIFIER' => '',
+            'PRICE_QUALIFIER' => '2', // 2 – Guide Price,
             'PROP_SUB_ID' => '0',
             'ADDRESS_1' => 'SZ ADF TestingEstate Agency. ZG Test (ML)',
             'ADDRESS_2' => 'Snowdon Drive, Winterhill',
@@ -68,6 +69,7 @@ class MediaColumnsMustAppearAfterAllOtherFieldsTest extends TestCase
             'FEATURE1' => 'House',
             'FEATURE2' => 'Garden',
             'FEATURE3' => 'Lake',
+            'LET_TYPE_ID' => '0', // 0 = not specified DEFAULT
             'MEDIA_IMAGE_00' => 'XX99XX_FBM2766_IMG_00.jpg',
             'SUMMARY' => 'whatever whatever whatever whatever',
             'DESCRIPTION' => 'whatever whatever whatever whatever whatever whatever',